services/agent-worker/requirements.txt
google-adk>=0.1.0
mcp>=1.0.0
PyGithub>=2.1.1
python-dotenv>=1.0.0
